Add web tool as a link at the top of the site
The cats-eye web tool is now included in the website, linked to at the
top of every page. The tool itself is just the application, so the
layout the rest of the site is not included.

The generated output is included (including the Gemfile dependency file)
because that seems to be the standard for this branch of the repository.